+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Added `/api/v2/filters` API endpoint returning empty array.
+- Support ~~strikethrough~~ text in posts.
 
 ### Changed